Greatly improved performance, fixed TAA NaNs, tweaked shadows, tweaked clouds & fog,...

+++ shaders/include/internalSettings.glsl
@@ -24,7 +24,7 @@ const int colortex8Format  = RGBA16F;   // History
 const int colortex9Format  = RG32UI;    // First bounce indirect & direct
 const int colortex10Format = RG16F;     // Variance, previous depth
 const int colortex11Format = RG32UI;    // Fog
-const int colortex12Format = RGBA32F;   // Ambient occlusion
+const int colortex12Format = RGBA16F;   // Ambient occlusion
 const int colortex15Format = RGBA16F;   // Gbuffer data
 
 const bool colortex0Clear  = false;
